write in memory raw data

+++ rover_system/driver/arch/pc/src/hc_sr04.c
@@ -10,7 +10,7 @@ int HC_SR04_init(void)
   return EXIT_SUCCESS;
 }
 
-int HC_SR04_getCM(void)
+float HC_SR04_getCM(void)
 {
   return (rand() % 100) + 50;
 }
